Move `TestCase` into the `ipl\Sql\Test` namespace
`TestCase` was defined in `tests/` which is not part of the distributed
library. Dependents such as `ipl-orm` had to work around this with a vendor
path mapping to `ipl-sql`'s test directory. Moving it into `src/Test/` makes it
a proper member of the `ipl\Sql\Test` testing utilities alongside
`SqlAssertions`, `TestAdapter`, and `TestConnection`.
