+++ tests/UpdateTest.php
@@ -4,6 +4,7 @@ namespace ipl\Tests\Sql;
 
 use ipl\Sql\Expression;
 use ipl\Sql\Select;
+use ipl\Sql\Test\TestCase;
 use ipl\Sql\Update;
 
 class UpdateTest extends TestCase
